requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# LangChain and AI dependencies - using more stable versions
langchain-community==0.0.10
langchain-groq==0.0.1
langchain-core==0.1.10

# Document processing
pypdf==3.17.4
python-docx==1.1.0
docx2txt==0.8

# Vector embeddings and storage - using lighter versions
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Additional utilities
requests==2.31.0
numpy==1.24.3
setuptools==68.2.2
